requirements.txt
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==8.0.0
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
pymessenger==0.0.7.0
pymessenger2==3.1.0
pyparsing==2.4.7
pytest==6.1.1
python-dotenv==0.14.0
pyunsplash==1.0.0rc2
requests==2.24.0
requests-toolbelt==0.9.1
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
urllib3==1.25.10
Werkzeug==1.0.1
